refactor: strToInt 메서드를 Utils로 이동
