+++ src/test/java/racingcar/validator/MoveCountValidatorTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,19 +11,19 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static racingcar.validator.MoveCountValidator.*;
 
 class MoveCountValidatorTest {
-	@ParameterizedTest
-	@ValueSource(strings = {"a", " ", "가"})
-	@DisplayName("입력값이 숫자가 아닐 경우")
-	void notInteger(String moveCount) {
-		assertThatThrownBy(() -> MoveCountValidator.validateMoveCount(moveCount))
-			.isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining(moveCount);
-	}
+	// @ParameterizedTest
+	// @ValueSource(strings = {"a", " ", "가"})
+	// @DisplayName("입력값이 숫자가 아닐 경우")
+	// void notInteger(String moveCount) {
+	// 	assertThatThrownBy(() -> MoveCountValidator.validateMoveCount(moveCount))
+	// 		.isInstanceOf(IllegalArgumentException.class)
+	// 		.hasMessageContaining(ERROR_NOT_INTEGER);
+	// }
 
 	@ParameterizedTest
-	@ValueSource(strings = {"-1", "0"})
+	@ValueSource(ints = {-1, 0})
 	@DisplayName("입력값이 양수가 아닐 경우")
-	void notPositive(String moveCount) {
+	void notPositive(int moveCount) {
 		assertThatThrownBy(() -> MoveCountValidator.validateMoveCount(moveCount))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining(ERROR_NOT_POSITIVE);
